remove old files and reorganize
