+++ mealie/db/sql/model_base.py
@@ -1,3 +1,5 @@
+from typing import List
+
 import sqlalchemy.ext.declarative as dec
 
 SqlAlchemyBase = dec.declarative_base()
@@ -6,14 +8,12 @@ SqlAlchemyBase = dec.declarative_base()
 class BaseMixins:
     @staticmethod
     def _sql_remove_list(session, list_of_tables: list, parent_id):
-        """
-        docstring
-        """
+
         for table in list_of_tables:
             session.query(table).filter_by(parent_id=parent_id).delete()
 
     @staticmethod
-    def _flatten_dict(list_of_dict: list[dict]):
+    def _flatten_dict(list_of_dict: List[dict]):
         finalMap = {}
         for d in list_of_dict:
