+++ mealie/db/sql/settings_models.py
@@ -35,6 +35,7 @@ class WebHookModel(SqlAlchemyBase, BaseMixins):
     def __init__(
         self, webhookURLs: list, webhookTime: str, enabled: bool = False
     ) -> None:
+
         self.webhookURLs = [WebhookURLModel(url=x) for x in webhookURLs]
         self.webhookTime = webhookTime
         self.enabled = enabled
@@ -42,6 +43,7 @@ class WebHookModel(SqlAlchemyBase, BaseMixins):
     def update(
         self, session, webhookURLs: list, webhookTime: str, enabled: bool
     ) -> None:
+    
         self._sql_remove_list(session, [WebhookURLModel], self.id)
 
         self.__init__(webhookURLs, webhookTime, enabled)
@@ -63,64 +65,3 @@ class WebhookURLModel(SqlAlchemyBase):
 
     def to_str(self):
         return self.url
-
-
-class SiteThemeModel(SqlAlchemyBase):
-    __tablename__ = "site_theme"
-    name = sa.Column(sa.String, primary_key=True)
-    colors = orm.relationship("ThemeColorsModel", uselist=False, cascade="all, delete")
-
-    def __init__(self, name: str, colors: dict) -> None:
-        self.name = name
-        self.colors = ThemeColorsModel(**colors)
-
-    def update(self, name, colors: dict) -> dict:
-        self.colors.update(**colors)
-        return self.dict()
-
-    def dict(self):
-        data = {"name": self.name, "colors": self.colors.dict()}
-        return data
-
-
-class ThemeColorsModel(SqlAlchemyBase):
-    __tablename__ = "theme_colors"
-    id = sa.Column(sa.Integer, primary_key=True)
-    parent_id = sa.Column(sa.String, sa.ForeignKey("site_theme.name"))
-    primary = sa.Column(sa.String)
-    accent = sa.Column(sa.String)
-    secondary = sa.Column(sa.String)
-    success = sa.Column(sa.String)
-    info = sa.Column(sa.String)
-    warning = sa.Column(sa.String)
-    error = sa.Column(sa.String)
-
-    def update(
-        self,
-        primary: str = None,
-        accent: str = None,
-        secondary: str = None,
-        success: str = None,
-        info: str = None,
-        warning: str = None,
-        error: str = None,
-    ) -> None:
-        self.primary = primary
-        self.accent = accent
-        self.secondary = secondary
-        self.success = success
-        self.info = info
-        self.warning = warning
-        self.error = error
-
-    def dict(self):
-        data = {
-            "primary": self.primary,
-            "accent": self.accent,
-            "secondary": self.secondary,
-            "success": self.success,
-            "info": self.info,
-            "warning": self.warning,
-            "error": self.error,
-        }
-        return data
