+++ pyproject.toml
@@ -25,6 +25,10 @@ tinydb = "^4.3.0"
 tinydb-serialization = "^2.0.0"
 python-slugify = "^4.0.1"
 requests = "^2.25.1"
+PyYAML = "^5.3.1"
+extruct = "^0.12.0"
+scrape-schema-recipe = "^0.1.3"
+python-multipart = "^0.0.5"
 
 [tool.poetry.dev-dependencies]
 pylint = "^2.6.0"
